+++ utils/themeeditor/quazip/zip.c
@@ -130,7 +130,7 @@ typedef struct
     int  encrypt;
 #ifndef NOCRYPT
     unsigned long keys[3];     /* keys defining the pseudo-random sequence */
-    const unsigned long* pcrc_32_tab;
+    const uint32_t* pcrc_32_tab;
     int crypt_header_size;
 #endif
 } curfile_info;
